Dynamic text analysis generation v1.0 implemented.
It took a long time for this to come together given the amount of frontend UI logic, adhering to type definitions (not used to that), and actually building the analysis. Right now it isn't the most accurate or interesting analysis (apart from the graphs), but I'm so glad that it finally works. Quite a few to-dos: fixing the readability metric data, improving the NLP analysis, adding PDF parsing, TXT document analysis (limited), markdown support (soon), analysis result logs, options to choose accuracy of analysis (stronger nlp models in the backend), etc.